Share pthreads code more explicitly in systhreads
Keep Win32 and Posix-specific functions in st_posix.h and st_win32.h
